+++ test/Tops/test_div_dipu.py
@@ -17,7 +17,7 @@ class MyModule(torch.nn.Module):
         r = torch.ops.aten.div(a, b)
         return r
 
-a = torch.arange(16).reshape(4, 4)
+a = torch.arange(16).reshape(4, 4).float()
 b = 1
 
 m = MyModule()
